🎉 FEAT(마일리지): 마일리지 차감, 되돌리기
결제 성공 후 마일리지 차감하기, 환불 후 마일리지 되돌리기를 추가했습니다

+++ src/main/java/com/supernova/fashionnova/domain/mileage/dto/MileageResponseDto.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 @Getter
 public class MileageResponseDto {
 
-    private final int usedMileage;
+    private final Long usedMileage;
     private final LocalDateTime createdAt;
 
     public MileageResponseDto(Order order) {
